pyproject.toml
[project]
name = "summary-compare"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.3.0",
    "datasets>=3.2.0",
    "python-dotenv>=1.0.1",
    "torch>=2.6.0",
    "transformers>=4.48.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.4",
]
